pyproject.toml
[tool.poetry]
name = "protowrap"
version = "0.0.1-pre"
description = "proto wrap api"
readme = "README.md"
authors = [
    "Matthew Taylor"
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
toml = "^0.10.2"
grpcio = "^1.73.0"
grpcio-tools = "^1.73.0"
grpclib = "^0.4.7"
pillow = "10.4.0"
matplotlib = "3.9.2"
pandas = "2.2.3"
betterproto = { version = "2.0.0b6", extras = ["compiler"] }